Add files via upload
This code is used to calculate the A/G factor, the equivalent uniform annual series, of a given problem. Unlike the F/P and P/F factor codes, only P and G can be unknown variables. If you have a method to generate formulas to solve i and n, respectively, you can make a pull request or fork the code. I would appreciate any contributions!